Small refactor #3
