+++ src/main/java/com/codecool/category/CategoryConverter.java
@@ -1,7 +1,7 @@
-package com.codecool.converter;
+package com.codecool.category;
 
 import com.codecool.category.domain.Category;
-import com.codecool.modelDTO.CategoryDTO;
+import com.codecool.category.dto.CategoryDTO;
 import org.springframework.stereotype.Component;
 import java.util.List;
 import java.util.stream.Collectors;
